insert/update nominativo

+++ academy-fullstack-bean/src/main/java/it/previnet/academy/fullstack/bean/intf/WithToken.java
@@ -2,5 +2,5 @@ package it.previnet.academy.fullstack.bean.intf;
 
 public interface WithToken {
 
-    Long getToken();
+    Integer getToken();
 }
